Dev/review service

+++ review-service/src/main/java/com/aresky/reviewservice/entity/ReviewStatistic.java
@@ -23,5 +23,5 @@ public class ReviewStatistic {
     private Integer numberOfReviews;
 
     @Column("stars_average")
-    private Integer starsAverage;
+    private Float starsAverage;
 }
